Memory Serialization & bug fixes
Added new SporeMemory class which handles the global memory via the RawMemory API. Currently disabled from using fast-stringify due to the unpackaged debug builds not including node modules when uploaded. Didn't see a notable improvement switching serializers, or from caching the global state to reduce parses actually...

Fixed a bug in the SporePathFinder which would throw an exception if a creeps starting path was on a room boundary.